+++ CHANGELOG.md
@@ -7,10 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+## [0.2.4] - rc (2017-04-16)
+
 - add pretty error handling to the command registrar
 - Move the writing-to-fs functionality to [bit-scope-client](https://github.com/teambit/bit-scope-client) project
 - Add opts to the load function
 - Add resolve function
 - Fix dist bug
 - add sourcemaps
-- update importer to version 0.5.0
+- update bit-scope-client to version 0.5.2